+++ src/main/java/com/osmimport/input/csv/ParserCallBack.java
@@ -4,7 +4,7 @@ import com.osmimport.model.OSMAttributedEntity;
 
 public interface ParserCallBack {
 
-	public void lineParsed(long lineNumber, OSMAttributedEntity entity);
+	public void lineParsed(long lineNumber, OSMAttributedEntity entity) throws Exception ;
 	
 	public void invalidLine(long lineNumber, String line);