+++ src/main/java/com/osmimport/parsing/csv/FolderCSVInput.java
@@ -1,0 +1,17 @@
+package com.osmimport.parsing.csv;
+
+import java.io.File;
+
+/**
+ * this is a parsing subsystem, using a folder and csv files
+ * 
+ * @author pfreydiere
+ * 
+ */
+public class FolderCSVInput {
+
+	public FolderCSVInput(File folder) {
+		
+	}
+	
+}